third_party/WebKit/LayoutTests/platform/win/external/wpt/user-timing/test_user_timing_clear_measures-expected.txt

This is a testharness.js-based test.
PASS window.performance is defined 
PASS Two measures have been created for this test. 
PASS After a call to window.performance.clearMeasures("measureUndefined"), where "measureUndefined" is a non-existent measure, window.performance.getEntriesByName("measure1") returns an object containing the "measure1" measure. 
PASS After a call to window.performance.clearMeasures("measureUndefined"), where "measureUndefined" is a non-existent measure, window.performance.getEntriesByName("measure2") returns an object containing the "measure2" measure. 
PASS After a call to window.performance.clearMeasures("measure1"), window.performance.getEntriesByName("measure1") returns an empty object. 
PASS After a call to window.performance.clearMeasures("measure1"), window.performance.getEntriesByName("measure2") returns an object containing the "measure2" measure. 
PASS After a call to window.performance.clearMeasures(), window.performance.getEntriesByType("measure") returns an empty object. 
Harness: the test ran to completion.
